new_version/new-grpc-api/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
grpcio-reflection==1.60.0
protobuf==4.25.1
psycopg==3.1.13
python-dotenv==1.0.0
